pyhton-requirements.txt
absl-py==1.0.0
anyio==3.6.1
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.1
asttokens==2.0.5
astunparse==1.6.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
black==22.3.0
bleach==5.0.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.2.0
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.30
dbus-python==1.2.18
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
deluge==2.0.3
distro==1.7.0
distro-info===1.1build1
Django==4.0.4
djangorestframework==3.13.1
dnspython==2.1.0
docopt==0.6.2
duplicity==0.8.21
entrypoints==0.4
executing==0.8.3
fasteners==0.14.1
fastjsonschema==2.15.3
flatbuffers==1.12
fonttools==4.33.3
future==0.18.2
gast==0.4.0
GeoIP==1.3.2
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h5py==3.7.0
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
imageio==2.19.1
importlib-metadata==4.6.4
incremental==21.3.0
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
isodate==0.6.1
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.8
jsonschema==4.6.0
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==3.4.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.4.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.1
libtorrent===2.0.5-build-libtorrent-rasterbar-qrM5vM-libtorrent-rasterbar-2.0.5-bindings-python
llvmlite==0.38.1
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mediapipe==0.8.9.1
mistune==0.8.4
molesq==0.4.0
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==0.4.3
nbclassic==0.3.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.11.0
networkx==2.8
nltk==3.7
notebook==6.4.11
notebook-shim==0.1.0
numba==0.55.2
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
opencv-python==4.5.5.64
opt-einsum==3.3.0
Owlready2==0.37
packaging==21.3
pandas==1.4.2
pandoc==2.2
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.2
plotly==5.8.2
plumbum==1.7.2
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.12.4
proton-client==0.7.1
protonvpn-cli==3.12.0
protonvpn-nm-lib==3.11.0
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pygame==2.1.2
Pygments==2.11.2
PyGObject==3.42.0
PyHamcrest==2.0.2
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.1.1
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt==5.15.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-gnupg==0.4.8
pythondialog==3.5.1
pytz==2022.1
PyWavelets==1.3.0
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==23.1.0
rdflib==6.1.1
regex==2022.6.2
rencode==1.0.6
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
scikit-fuzzy==0.4.2
scikit-image==0.19.2
scikit-learn==1.0.2
scikit-surprise==1.1.1
scipy==1.8.0
scour==0.38.2
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==3.3.1
Send2Trash==1.8.0
service-identity==18.1.0
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
systemd-python==234
tenacity==8.0.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.5.4
tinycss2==1.1.1
tk==0.1.0
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
Twisted==22.1.0
typing_extensions==4.2.0
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
variety==0.8.5
vboxapi==1.0
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.2
wrapt==1.14.1
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
zope.interface==5.4.0
